# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MEDI-IOT EyeCare API",
    "description": "\n## MEDI-IOT EyeCare — 안과 AI 진단 플랫폼\n\n### 주요 기능\n- **환자 관리**: 등록, 조회, PII 암호화 저장\n- **검사 기록**: OCT, 안저, 시야 검사 등 다양한 검사 유형 지원\n- **AI 진단**: shared-libraries Orchestrator (CONSENSUS 전략) 기반\n  - FAST 모델: gemma-4-e4b (초안 생성)\n  - HEAVY 모델: gemma-4-26b-a4b (의료 정확성 검토)\n  - OntologyValidator: MEDICAL 도메인 규칙 자동 검증\n\n### 설계 원칙\n- 환자 PII는 암호화하여 저장 (응답 시 마스킹)\n- 의료 안전을 위해 CONSENSUS 전략 사용 (두 모델이 합의해야 PASS)\n- Circuit Breaker: 최대 2회 반복 후 현재 최선 결과 반환\n    ",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/diagnosis/exam": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "검사 기록 등록",
        "description": "안과 검사 기록을 등록합니다.\n\n등록 후 `POST /diagnosis/ai-analyze`를 호출하면 AI 진단 보고서가 생성됩니다.\n\n**지원 검사 종류**:\n| exam_type | 설명 |\n|-----------|------|\n| fundus | 안저 촬영 (당뇨망막병증, 녹내장 선별) |\n| oct | 빛간섭단층촬영 (황반 질환, 시신경 분석) |\n| visual_field | 시야 검사 (녹내장 진행 모니터링) |\n| slit_lamp | 세극등 검사 (각막, 수정체, 전방) |\n| refraction | 굴절 검사 (근시, 난시, 원시 측정) |\n| iop | 안압 검사 (녹내장 위험 평가) |",
        "operationId": "create_exam_api_v1_diagnosis_exam_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "등록된 검사 기록 (AI 분석 전 상태, report_status=pending)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "AI 진단 보고서 생성",
        "description": "검사 기록 기반으로 AI 진단 보고서를 생성합니다.\n\n**처리 파이프라인 (CONSENSUS 전략)**:\n```\n검사 소견 입력\n    ↓\nPlannerAgent (gemma-4-e4b)     ← 진단 작업 계획 수립\n    ↓\nGeneratorAgent (gemma-4-e4b)   ← 초안 보고서 생성\n    ↓\nReviewerAgent (gemma-4-26b-a4b) + OntologyValidator (MEDICAL)\n    ↓ PASS → 보고서 확정\n    ↓ FAIL → FixerAgent → 재검토 (최대 2회)\n    ↓\nDiagnosisResponse 반환\n```\n\n**예상 처리 시간**: 1~3분 (CONSENSUS 전략, LM Studio 기준)\n\n**OntologyValidator 검증 항목**:\n- Semantic: 의학 용어 정합성\n- Structural: 보고서 구조 완결성\n- Constraint: PII 미포함 여부\n- Dependency: ICD 코드 ↔ 진단명 일치\n\n> **주의**: `ontology_passed=false`인 경우 보고서가 반환되더라도\n> 의사의 추가 검토가 필요합니다.",
        "operationId": "create_diagnosis_api_v1_diagnosis__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "생성된 진단 보고서 (OntologyValidator 검증 완료)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/ai-analyze": {
      "post": {
        "tags": [
          "diagnosis"
        ],
        "summary": "AI 진단 보고서 생성 (명시적 경로)",
        "description": "검사 기록 기반으로 AI 진단 보고서를 생성합니다.\n\n**처리 파이프라인 (CONSENSUS 전략)**:\n```\n검사 소견 입력\n    ↓\nPlannerAgent (gemma-4-e4b)     ← 진단 작업 계획 수립\n    ↓\nGeneratorAgent (gemma-4-e4b)   ← 초안 보고서 생성\n    ↓\nReviewerAgent (gemma-4-26b-a4b) + OntologyValidator (MEDICAL)\n    ↓ PASS → 보고서 확정\n    ↓ FAIL → FixerAgent → 재검토 (최대 2회)\n    ↓\nDiagnosisResponse 반환\n```\n\n**예상 처리 시간**: 1~3분 (CONSENSUS 전략, LM Studio 기준)\n\n**OntologyValidator 검증 항목**:\n- Semantic: 의학 용어 정합성\n- Structural: 보고서 구조 완결성\n- Constraint: PII 미포함 여부\n- Dependency: ICD 코드 ↔ 진단명 일치\n\n> **주의**: `ontology_passed=false`인 경우 보고서가 반환되더라도\n> 의사의 추가 검토가 필요합니다.",
        "operationId": "create_diagnosis_api_v1_diagnosis_ai_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "생성된 진단 보고서 (OntologyValidator 검증 완료)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosis/{diagnosis_id}": {
      "get": {
        "tags": [
          "diagnosis"
        ],
        "summary": "진단 결과 조회",
        "description": "진단 결과를 조회합니다.",
        "operationId": "get_diagnosis_api_v1_diagnosis__diagnosis_id__get",
        "parameters": [
          {
            "name": "diagnosis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Diagnosis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DiagnosisRequest": {
        "properties": {
          "exam_id": {
            "type": "string",
            "title": "Exam Id",
            "description": "AI 분석할 검사 기록 UUID (eye_exams.id)"
          },
          "additional_context": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Additional Context",
            "description": "추가 임상 맥락 (혈당 수치, 복약 이력 등 — AI 정확도 향상에 도움)"
          },
          "strategy": {
            "type": "string",
            "pattern": "^(pipeline|consensus|debate|fastest)$",
            "title": "Strategy",
            "description": "Orchestrator 전략:\n- **consensus** (권장): FAST+HEAVY 두 모델 합의 — 의료 안전 최적\n- pipeline: 순차 실행 (Planner→Generator→Reviewer→Fixer)\n- debate: 두 모델 논쟁 후 합의\n- fastest: 가장 빠른 모델 단독 실행 (응급 상황)",
            "default": "consensus",
            "examples": [
              "consensus"
            ]
          }
        },
        "type": "object",
        "required": [
          "exam_id"
        ],
        "title": "DiagnosisRequest",
        "example": {
          "additional_context": "환자 HbA1c 8.2%, 당뇨병 진단 12년차, 인슐린 치료 중",
          "exam_id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
          "strategy": "consensus"
        }
      },
      "DiagnosisResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "진단 UUID"
          },
          "exam_id": {
            "type": "string",
            "title": "Exam Id",
            "description": "관련 검사 UUID"
          },
          "diagnosis_code": {
            "type": "string",
            "title": "Diagnosis Code",
            "description": "ICD-10 진단 코드 (예: H36.0)"
          },
          "diagnosis_name": {
            "type": "string",
            "title": "Diagnosis Name",
            "description": "진단명 (한국어)"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "중증도: normal | mild | moderate | severe | critical"
          },
          "report": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report",
            "description": "LLM 생성 진단 보고서 전문 (OntologyValidator 검증 완료, 최대 2000자)"
          },
          "treatment_plan": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Treatment Plan",
            "description": "치료 계획 및 추적 관찰 권고"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model",
            "description": "보고서 생성에 사용된 LLM 모델명"
          },
          "llm_iterations": {
            "type": "integer",
            "title": "Llm Iterations",
            "description": "Orchestrator 반복 횟수 (1 = 첫 시도 통과)"
          },
          "llm_latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Latency Ms",
            "description": "LLM 처리 시간 (밀리초)"
          },
          "ontology_passed": {
            "type": "boolean",
            "title": "Ontology Passed",
            "description": "OntologyValidator MEDICAL 도메인 검증 통과 여부"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score",
            "description": "진단 신뢰도 (0.0~1.0, ontology_passed=True면 0.85+)"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "생성일시 (UTC)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "exam_id",
          "diagnosis_code",
          "diagnosis_name",
          "severity",
          "report",
          "treatment_plan",
          "llm_model",
          "llm_iterations",
          "llm_latency_ms",
          "ontology_passed",
          "confidence_score",
          "created_at"
        ],
        "title": "DiagnosisResponse",
        "example": {
          "confidence_score": 0.85,
          "created_at": "2026-05-09T10:40:00Z",
          "diagnosis_code": "H36.0",
          "diagnosis_name": "당뇨망막병증",
          "exam_id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
          "id": "c3d4e5f6-a7b8-9012-cdef-123456789012",
          "llm_iterations": 1,
          "llm_latency_ms": 85420.5,
          "llm_model": "google/gemma-4-26b-a4b",
          "ontology_passed": true,
          "report": "## 안과 진단 보고서\n\n**진단**: 비증식성 당뇨망막병증 중등도 (H36.0)\n\n**소견 요약**: 우안에서 황반 주위 점상출혈 및 경성삼출물이 다수 관찰되며, 신생혈관 의심 소견이 있습니다. 좌안은 경미한 미세동맥류 소견입니다.\n\n**중증도**: 중등도 — 즉각적인 안과 전문의 추적 관찰 필요\n\n**치료 권고**: 범망막광응고술(PRP) 시술 고려, 내분비과 협진을 통한 혈당 조절 강화",
          "severity": "moderate",
          "treatment_plan": "3개월 내 범망막광응고술(PRP) 시술 검토. 혈당 목표 HbA1c < 7%."
        }
      },
      "ExamCreate": {
        "properties": {
          "patient_id": {
            "type": "string",
            "title": "Patient Id",
            "description": "환자 UUID (patients.id)"
          },
          "exam_type": {
            "type": "string",
            "pattern": "^(fundus|oct|visual_field|slit_lamp|refraction|iop)$",
            "title": "Exam Type",
            "description": "검사 종류:\n- fundus: 안저 촬영\n- oct: 빛간섭단층촬영\n- visual_field: 시야 검사\n- slit_lamp: 세극등 검사\n- refraction: 굴절 검사\n- iop: 안압 검사",
            "examples": [
              "fundus"
            ]
          },
          "exam_date": {
            "type": "string",
            "format": "date",
            "title": "Exam Date",
            "description": "검사 날짜 (YYYY-MM-DD)"
          },
          "icd_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icd Code",
            "description": "ICD-10 코드 (예: H36.0, H40.1)",
            "examples": [
              "H36.0"
            ]
          },
          "iop_left": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 80.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Iop Left",
            "description": "좌안 안압 (mmHg, 정상 범위: 10~21)"
          },
          "iop_right": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 80.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Iop Right",
            "description": "우안 안압 (mmHg)"
          },
          "visual_acuity_left": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Acuity Left",
            "description": "좌안 교정시력 (예: 0.8, 1.0, 0.1)"
          },
          "visual_acuity_right": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Acuity Right",
            "description": "우안 교정시력"
          },
          "raw_findings": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Findings",
            "description": "의사가 직접 입력한 검사 소견 원문 (AI 분석 기반 데이터)"
          }
        },
        "type": "object",
        "required": [
          "patient_id",
          "exam_type",
          "exam_date"
        ],
        "title": "ExamCreate",
        "example": {
          "exam_date": "2026-05-09",
          "exam_type": "fundus",
          "icd_code": "H36.0",
          "iop_left": 14.5,
          "iop_right": 15.2,
          "patient_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "raw_findings": "우안: 황반 주위 점상출혈 및 경성삼출물 다수 관찰. 신생혈관 의심 소견. 좌안: 경미한 미세동맥류 관찰.",
          "visual_acuity_left": "0.8",
          "visual_acuity_right": "0.7"
        }
      },
      "ExamResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "검사 UUID"
          },
          "patient_id": {
            "type": "string",
            "title": "Patient Id",
            "description": "환자 UUID"
          },
          "exam_type": {
            "type": "string",
            "title": "Exam Type",
            "description": "검사 종류"
          },
          "exam_date": {
            "type": "string",
            "format": "date",
            "title": "Exam Date",
            "description": "검사 날짜"
          },
          "icd_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icd Code",
            "description": "ICD-10 코드"
          },
          "iop_left": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iop Left",
            "description": "좌안 안압 (mmHg)"
          },
          "iop_right": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iop Right",
            "description": "우안 안압 (mmHg)"
          },
          "visual_acuity_left": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Acuity Left",
            "description": "좌안 시력"
          },
          "visual_acuity_right": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Acuity Right",
            "description": "우안 시력"
          },
          "raw_findings": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Findings",
            "description": "검사 소견 원문"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary",
            "description": "LLM 생성 요약 (최대 500자)"
          },
          "report_status": {
            "type": "string",
            "title": "Report Status",
            "description": "보고서 상태: pending | generating | completed | failed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "등록일시 (UTC)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "exam_type",
          "exam_date",
          "icd_code",
          "iop_left",
          "iop_right",
          "visual_acuity_left",
          "visual_acuity_right",
          "raw_findings",
          "ai_summary",
          "report_status",
          "created_at"
        ],
        "title": "ExamResponse",
        "example": {
          "ai_summary": "당뇨망막병증 비증식성 중등도 소견. 추가 OCT 권장.",
          "created_at": "2026-05-09T10:35:00Z",
          "exam_date": "2026-05-09",
          "exam_type": "fundus",
          "icd_code": "H36.0",
          "id": "b2c3d4e5-f6a7-8901-bcde-f12345678901",
          "iop_left": 14.5,
          "iop_right": 15.2,
          "patient_id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
          "raw_findings": "우안: 황반 주위 점상출혈 및 경성삼출물 다수 관찰.",
          "report_status": "completed",
          "visual_acuity_left": "0.8",
          "visual_acuity_right": "0.7"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}